FINISHED THE LEDGE HANG ITS SO SMMOOTH
very happy with what i got done, the player movement pretty much complete which means we can actually start to work on the levels. Still need to finish the npc system and the text stuff as well as a few more obstacles but this is a big step in the right direction!